feat!: drop eslint, rename to oxlint-config-decent, adopt pnpm + oxc toolchain
BREAKING CHANGE: The ESLint config export has been removed and the package is
now published as oxlint-config-decent. The oxlint config is the root export;
the ./oxlint subpath remains as a compatibility alias.

- Remove the ESLint config (typescript-eslint caps TypeScript at <6.1.0,
  blocking TypeScript 7); the package is now oxlint-only with zero runtime
  dependencies
- Migrate npm -> pnpm 11 (pnpm-lock.yaml, pnpm-workspace.yaml, packageManager
  pin, .pnpmfile.cjs for the JS-API TypeScript 6 shim the compat plugins need)
- Replace unbuild with tsdown (dts generated by the native tsgo backend),
  the rules-parity script with vitest, prettier with oxfmt, and self-lint
  with oxlint dogfooding its own config
- Move to TypeScript 7.0.2; CI matrix now Node 22/24/26
- Release workflow uses a GitHub App token so semantic-release keeps pushing
  CHANGELOG.md updates to the protected main branch; drop npm audit signatures
  (near-zero coverage under pnpm's store layout); release from beta branch too
