+++ .github/renovate.json5
@@ -26,6 +26,12 @@
       groupName: 'dev dependencies',
       semanticCommitType: 'chore',
     },
+    // pnpm itself, pinned via the packageManager field (pnpm-lock.yaml is maintained natively).
+    {
+      matchDepTypes: ['packageManager'],
+      groupName: 'dev dependencies',
+      semanticCommitType: 'chore',
+    },
     {
       matchDepTypes: ['dependencies'],
       groupName: 'all dependencies',
