+++ .github/workflows/lint.yml
@@ -48,6 +48,9 @@ jobs:
           VALIDATE_JSONC: false
           VALIDATE_JSONC_PRETTIER: false
           VALIDATE_JSON_PRETTIER: false
+          # Markdown/YAML are formatted by oxfmt now; prettier would disagree.
+          VALIDATE_MARKDOWN_PRETTIER: false
+          VALIDATE_YAML_PRETTIER: false
           VALIDATE_JAVASCRIPT_ES: false
           VALIDATE_JAVASCRIPT_PRETTIER: false
           VALIDATE_JAVASCRIPT_STANDARD: false
