--- .npmignore
@@ -1,38 +1,0 @@
-# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Runtime data
-pids
-*.pid
-*.seed
-*.pid.lock
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-src/
-tests/
-node_modules/
-.editorconfig
-.gitignore
-.git
-.npmrc
-
-release.config.mjs
-tsconfig.json
-tsconfig.lint.json
-
-*~
-*#
-.DS_STORE
-.netbeans
-nbproject
-.idea
-.vscode
