--- eslint.config.mjs
@@ -1,5 +1,0 @@
-import { config } from './dist/index.mjs';
-
-export default config({
-  tsconfigRootDir: import.meta.dirname,
-});
