+++ package.json
@@ -1,11 +1,12 @@
 {
-  "name": "eslint-config-decent",
+  "name": "oxlint-config-decent",
   "version": "4.2.38",
-  "description": "A decent ESLint configuration",
+  "description": "A decent oxlint configuration for TypeScript projects",
   "main": "./dist/index.cjs",
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
   "type": "module",
+  "packageManager": "pnpm@11.11.0",
   "exports": {
     ".": {
       "import": {
@@ -30,16 +31,17 @@
   },
   "files": [
     "dist",
-    "src"
+    "src",
+    "skills"
   ],
   "scripts": {
-    "build": "unbuild",
-    "test": "tsgo --noEmit --skipLibCheck",
-    "test:oxlint-rules": "node scripts/test-oxlint-rules.mjs",
-    "lint:markdown": "prettier --write '**/*.md' '!**/node_modules/**' '!**/dist/**' && markdownlint '**/*.md' --ignore '**/node_modules/**' --ignore '**/dist/**' --config=.github/linters/.markdown-lint.yml --fix",
-    "lint:code": "eslint --fix",
-    "prelint": "npm run build",
-    "lint": "run-p lint:*",
+    "build": "tsdown",
+    "check:types": "tsc --noEmit",
+    "test": "pnpm run check:types && vitest run",
+    "lint:fmt": "oxfmt",
+    "lint:code": "oxlint --fix",
+    "lint:markdown": "markdownlint '**/*.md' --ignore '**/node_modules/**' --ignore '**/dist/**' --config=.github/linters/.markdown-lint.yml --fix",
+    "lint": "pnpm run /^lint:/",
     "lint-staged": "lint-staged",
     "beta": "npm publish --tag beta",
     "prepublishOnly": "pinst --disable",
@@ -47,151 +49,95 @@
     "prepare": "husky"
   },
   "lint-staged": {
+    "*.{js,cjs,mjs,jsx,ts,cts,mts,tsx}": [
+      "oxfmt",
+      "oxlint --fix"
+    ],
     "*.md": [
-      "prettier --write --cache",
+      "oxfmt",
       "markdownlint --config=.github/linters/.markdown-lint.yml --fix"
     ],
-    "*.{json5,yml}": [
-      "prettier --write"
-    ],
-    "*.{js,cjs,mjs,ts}": [
-      "eslint --fix"
+    "*.{json,jsonc,json5,yml,yaml}": [
+      "oxfmt"
     ]
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/jgeurts/eslint-config-decent.git"
+    "url": "git+https://github.com/jgeurts/oxlint-config-decent.git"
   },
   "keywords": [
-    "eslint",
     "oxlint",
     "config",
-    "decent"
+    "decent",
+    "typescript",
+    "linter"
   ],
   "author": "Jim Geurts <jim@biacreations.com>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/jgeurts/eslint-config-decent/issues"
+    "url": "https://github.com/jgeurts/oxlint-config-decent/issues"
   },
-  "homepage": "https://github.com/jgeurts/eslint-config-decent#readme",
+  "homepage": "https://github.com/jgeurts/oxlint-config-decent#readme",
   "engines": {
     "node": ">=22.13.0"
   },
   "peerDependencies": {
-    "eslint": "^9.0.0 || ^10.0.0",
-    "oxlint": ">=1.53.0",
-    "oxlint-tsgolint": ">=0.1.0",
-    "oxlint-plugin-eslint": ">=1.52.0",
     "@stylistic/eslint-plugin": ">=5.0.0",
-    "@vitest/eslint-plugin": ">=1.0.0",
     "@typescript-eslint/eslint-plugin": ">=8.0.0",
+    "@vitest/eslint-plugin": ">=1.0.0",
     "eslint-plugin-jsdoc": ">=62.0.0",
     "eslint-plugin-react": ">=7.0.0",
     "eslint-plugin-security": ">=4.0.0",
     "eslint-plugin-testing-library": ">=7.0.0",
     "eslint-plugin-unicorn": ">=63.0.0",
+    "oxlint": ">=1.53.0",
+    "oxlint-plugin-eslint": ">=1.52.0",
+    "oxlint-tsgolint": ">=0.1.0",
     "typescript": ">=5.5.0"
   },
   "peerDependenciesMeta": {
-    "oxlint": {
-      "optional": true
-    },
-    "oxlint-tsgolint": {
-      "optional": true
-    },
-    "oxlint-plugin-eslint": {
-      "optional": true
-    },
-    "@typescript-eslint/eslint-plugin": {
-      "optional": true
-    },
-    "@stylistic/eslint-plugin": {
-      "optional": true
-    },
     "@vitest/eslint-plugin": {
-      "optional": true
-    },
-    "eslint-plugin-jsdoc": {
       "optional": true
     },
     "eslint-plugin-react": {
       "optional": true
     },
-    "eslint-plugin-security": {
-      "optional": true
-    },
     "eslint-plugin-testing-library": {
       "optional": true
     },
-    "eslint-plugin-unicorn": {
+    "oxlint-tsgolint": {
+      "optional": true
+    },
+    "typescript": {
       "optional": true
     }
   },
-  "dependencies": {
-    "@eslint/compat": "2.1.0",
-    "@eslint/js": "10.0.1",
-    "@next/eslint-plugin-next": "16.2.10",
-    "@stylistic/eslint-plugin": "5.10.0",
-    "@vitest/eslint-plugin": "1.6.22",
-    "eslint-config-prettier": "10.1.8",
-    "eslint-plugin-import-x": "4.17.1",
-    "eslint-plugin-jsdoc": "63.0.12",
-    "eslint-plugin-jsx-a11y": "6.10.2",
-    "eslint-plugin-prettier": "5.5.6",
-    "eslint-plugin-promise": "7.3.0",
-    "eslint-plugin-react": "7.37.5",
-    "eslint-plugin-react-hooks": "7.1.1",
-    "eslint-plugin-security": "4.0.1",
-    "eslint-plugin-testing-library": "7.16.2",
-    "eslint-plugin-unicorn": "71.1.0",
-    "globals": "17.7.0",
-    "typescript-eslint": "8.63.0"
-  },
   "devDependencies": {
-    "@eslint/config-helpers": "0.6.0",
     "@semantic-release/changelog": "6.0.3",
     "@semantic-release/commit-analyzer": "13.0.1",
     "@semantic-release/git": "10.0.1",
     "@semantic-release/github": "12.0.9",
     "@semantic-release/npm": "13.1.5",
     "@semantic-release/release-notes-generator": "14.1.1",
-    "@swc/core": "1.15.43",
+    "@stylistic/eslint-plugin": "5.10.0",
     "@types/node": ">=22",
-    "@typescript/native-preview": "7.0.0-dev.20260707.2",
+    "@typescript-eslint/eslint-plugin": "8.63.0",
+    "@vitest/eslint-plugin": "1.6.21",
     "conventional-changelog-conventionalcommits": "10.2.1",
-    "eslint": "10.6.0",
+    "eslint-plugin-jsdoc": "63.0.12",
+    "eslint-plugin-security": "4.0.1",
+    "eslint-plugin-unicorn": "71.1.0",
     "husky": "9.1.7",
     "lint-staged": "17.0.8",
     "markdownlint-cli": "0.49.0",
-    "npm-run-all2": "9.0.2",
+    "oxfmt": "0.58.0",
     "oxlint": "1.73.0",
+    "oxlint-plugin-eslint": "1.73.0",
+    "oxlint-tsgolint": "0.24.0",
     "pinst": "3.0.0",
-    "prettier": "3.9.5",
-    "semantic-release": "25.0.5",
-    "typescript": "6.0.3",
-    "unbuild": "3.6.1"
-  },
-  "overrides": {
-    "eslint-plugin-import-x": {
-      "eslint": "$eslint"
-    },
-    "eslint-plugin-jsx-a11y": {
-      "eslint": "$eslint"
-    },
-    "eslint-plugin-promise": {
-      "eslint": "$eslint"
-    },
-    "eslint-plugin-react": {
-      "eslint": "$eslint"
-    },
-    "eslint-plugin-react-hooks": {
-      "eslint": "$eslint"
-    },
-    "eslint-plugin-testing-library": {
-      "eslint": "$eslint"
-    },
-    "unbuild": {
-      "typescript": "$typescript"
-    }
+    "semantic-release": "25.0.6",
+    "tsdown": "0.22.4",
+    "typescript": "7.0.2",
+    "vitest": "4.1.10"
   }
 }
