+++ release.config.mjs
@@ -46,7 +46,7 @@ export default {
     [
       '@semantic-release/git',
       {
-        assets: ['CHANGELOG.md', 'package.json', 'package-lock.json'],
+        assets: ['CHANGELOG.md', 'package.json'],
         message: `chore: release \${nextRelease.version} [skip ci]\n\n\${nextRelease.notes}`,
       },
     ],
