--- src/types/eslint-plugin-jsx-a11y.d.ts
@@ -1,10 +1,0 @@
-declare module 'eslint-plugin-jsx-a11y' {
-  import { type Config, type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin & {
-    configs: {
-      recommended: Config;
-    };
-  };
-  export default value;
-}
