--- src/types/eslint-plugin-next.d.ts
@@ -1,11 +1,0 @@
-declare module '@next/eslint-plugin-next' {
-  import { type Config, type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin & {
-    configs: {
-      recommended: Config;
-      'core-web-vitals': Config;
-    };
-  };
-  export default value;
-}
