--- src/types/eslint-plugin-promise.d.ts
@@ -1,6 +1,0 @@
-declare module 'eslint-plugin-promise' {
-  import { type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin;
-  export default value;
-}
