--- src/types/eslint-plugin-react-hooks.d.ts
@@ -1,12 +1,0 @@
-declare module 'eslint-plugin-react-hooks' {
-  import { type Config, type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin & {
-    configs: {
-      flat: {
-        'recommended-latest': Config;
-      };
-    };
-  };
-  export default value;
-}
