--- src/types/eslint-plugin-security.d.ts
@@ -1,6 +1,0 @@
-declare module 'eslint-plugin-security' {
-  import { type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin;
-  export default value;
-}
