--- src/types/eslint-plugin-testing-library.d.ts
@@ -1,13 +1,0 @@
-declare module 'eslint-plugin-testing-library' {
-  import { type Config, type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin & {
-    configs: {
-      'flat/angular': Config;
-      'flat/dom': Config;
-      'flat/react': Config;
-      'flat/vue': Config;
-    };
-  };
-  export default value;
-}
