--- src/types/eslint-plugin-unicorn.d.ts
@@ -1,6 +1,0 @@
-declare module 'eslint-plugin-unicorn' {
-  import { type Plugin } from '@eslint/config-helpers';
-
-  const value: Plugin;
-  export default value;
-}
