--- src/unicorn.ts
@@ -1,26 +1,0 @@
-import { type Config } from '@eslint/config-helpers';
-import unicorn from 'eslint-plugin-unicorn';
-
-const base: Config = {
-  plugins: {
-    unicorn,
-  },
-  rules: {
-    'unicorn/better-regex': 'error',
-    'unicorn/custom-error-definition': 'error',
-    'unicorn/no-array-method-this-argument': 'error',
-    'unicorn/no-for-loop': 'error',
-    'unicorn/prefer-array-find': 'error',
-    'unicorn/prefer-node-protocol': 'error',
-    'unicorn/prefer-object-from-entries': 'error',
-    'unicorn/prefer-set-has': 'error',
-  },
-};
-
-export const configs = {
-  base,
-};
-
-export default {
-  configs,
-};
