+++ tsconfig.json
@@ -6,12 +6,9 @@
     "moduleResolution": "node16",
     "types": ["node"],
     "allowJs": false,
-
-    "outDir": "dist",
-
-    "declaration": true,
-    "sourceMap": true,
-    "sourceRoot": ".",
+    "skipLibCheck": true,
+    "noEmit": true,
+    "allowImportingTsExtensions": true,
 
     "strict": true,
     "forceConsistentCasingInFileNames": true,
@@ -24,12 +21,6 @@
     "noUncheckedIndexedAccess": true,
     "useUnknownInCatchVariables": false
   },
-  "include": [
-    "src/**/*.ts",
-    "build.config.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
+  "include": ["src/**/*.ts", "tests/**/*.ts", "oxlint.config.ts", "tsdown.config.ts"],
+  "exclude": ["node_modules", "dist"]
 }
